requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
coverage==7.2.7
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
